@@ -1,115 +1,139 @@
 {
   "name": "itinerate-consulting",
-  "description": "The website for Itinerate Consulting, a micro-consultancy, specialising in web applications & JavaScript",
-  "keywords": [
-    "node",
-    "heroku",
-    "express"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/gazzwi86/itinerate-consulting"
-  },
-  "version": "0.0.1",
-  "license": "MIT",
-  "main": "server/app.js",
+  "version": "0.0.0",
+  "main": "server/index.js",
   "dependencies": {
-    "babel-runtime": "^5.8.20",
-    "bluebird": "^2.9.34",
-    "body-parser": "^1.13.3",
-    "composable-middleware": "^0.3.0",
-    "compression": "^1.5.2",
-    "connect-mongo": "^0.8.1",
-    "cookie-parser": "^1.3.5",
-    "ejs": "^2.3.3",
-    "errorhandler": "^1.4.2",
+    "angular": "~1.5.5",
+    "bootstrap-sass": "~3.3.7",
+    "bootstrap": "~3.3.7",
+    "angular-animate": "~1.5.5",
+    "angular-aria": "~1.5.5",
+    "angular-resource": "~1.5.5",
+    "angular-cookies": "~1.5.5",
+    "angular-sanitize": "~1.5.5",
+    "angular-route": "~1.5.5",
+    "angular-ui-bootstrap": "^2.0.1",
+    "font-awesome": ">=4.1.0",
+    "core-js": "^2.2.1",
     "express": "^4.13.3",
-    "express-session": "^1.11.3",
-    "kerberos": "0.0.17",
-    "lodash": "^3.10.1",
-    "lusca": "^1.3.0",
+    "morgan": "~1.7.0",
+    "body-parser": "^1.13.3",
     "method-override": "^2.3.5",
-    "mongoose": "^4.1.2",
-    "morgan": "~1.6.1",
-    "newrelic": "^1.24.0",
-    "serve-favicon": "^2.3.0"
+    "cookie-parser": "^1.3.5",
+    "express-session": "^1.11.3",
+    "errorhandler": "^1.4.2",
+    "composable-middleware": "^0.3.0",
+    "fast-json-patch": "^1.0.0",
+    "lodash": "^4.6.1",
+    "lusca": "^1.3.0",
+    "babel-runtime": "^6.6.1",
+    "babel-polyfill": "^6.7.2",
+    "pug": "2.0.0-beta4",
+    "serve-favicon": "^2.3.0",
+    "shrink-ray": "^0.1.3",
+    "sprint-js": "~0.1.0"
   },
   "devDependencies": {
+    "angular-mocks": "~1.5.5",
     "autoprefixer": "^6.0.0",
-    "babel-core": "^5.6.4",
-    "grunt": "~0.4.5",
-    "grunt-wiredep": "^2.0.0",
-    "grunt-concurrent": "^2.0.1",
-    "grunt-contrib-clean": "~0.7.0",
-    "grunt-contrib-concat": "^0.5.1",
-    "grunt-contrib-copy": "^0.8.0",
-    "grunt-contrib-cssmin": "~0.14.0",
-    "grunt-contrib-imagemin": "~1.0.0",
-    "grunt-contrib-jshint": "~0.11.2",
-    "grunt-contrib-uglify": "~0.11.0",
-    "grunt-contrib-watch": "~0.6.1",
-    "karma-babel-preprocessor": "^5.2.1",
-    "grunt-babel": "~5.0.0",
-    "grunt-google-cdn": "~0.4.0",
-    "grunt-jscs": "^2.1.0",
-    "grunt-newer": "^1.1.1",
-    "grunt-ng-annotate": "^1.0.1",
-    "grunt-ng-constant": "^1.1.0",
-    "grunt-filerev": "^2.3.1",
-    "grunt-usemin": "^3.0.0",
-    "grunt-env": "~0.4.1",
-    "grunt-node-inspector": "^0.4.1",
-    "grunt-nodemon": "^0.4.0",
-    "grunt-angular-templates": "^0.5.4",
-    "grunt-dom-munger": "^3.4.0",
-    "grunt-protractor-runner": "^2.0.0",
-    "grunt-injector": "^0.6.0",
-    "grunt-karma": "~0.12.0",
-    "grunt-build-control": "^0.6.0",
-    "grunt-contrib-sass": "^0.9.0",
-    "jit-grunt": "^0.9.1",
-    "time-grunt": "^1.2.1",
-    "grunt-express-server": "^0.5.1",
-    "grunt-postcss": "~0.7.1",
-    "grunt-open": "~0.2.3",
-    "open": "~0.0.4",
-    "jshint-stylish": "~2.1.0",
-    "connect-livereload": "^0.5.3",
-    "mocha": "^2.2.5",
-    "grunt-mocha-test": "~0.12.7",
-    "grunt-mocha-istanbul": "^3.0.1",
-    "istanbul": "~0.4.1",
+    "awesome-typescript-loader": "0.17.0",
+    "babel-core": "^6.6.5",
+    "babel-eslint": "^6.0.4",
+    "babel-loader": "^6.2.4",
+    "babel-plugin-syntax-flow": "^6.8.0",
+    "babel-plugin-transform-class-properties": "^6.6.0",
+    "babel-plugin-transform-flow-comments": "^6.8.0",
+    "babel-plugin-transform-runtime": "^6.6.0",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-register": "^6.6.5",
+    "browser-sync": "^2.8.0",
+    "bs-fullscreen-message": "^1.0.0",
     "chai": "^3.2.0",
-    "sinon": "^1.16.1",
     "chai-as-promised": "^5.1.0",
     "chai-things": "^0.2.0",
-    "sinon-chai": "^2.8.0",
+    "css-loader": "^0.24.0",
+    "del": "^2.0.2",
+    "eslint": "^2.12.0",
+    "extract-text-webpack-plugin": "^1.0.1",
+    "file-loader": "^0.9.0",
+    "flow-bin": "^0.31.0",
+    "grunt": "^1.0.1",
+    "grunt-build-control": "^0.7.0",
+    "gulp": "^3.9.1",
+    "gulp-babel": "^6.1.2",
+    "gulp-env": "^0.4.0",
+    "gulp-eslint": "^2.0.0",
+    "gulp-imagemin": "^3.0.1",
+    "gulp-inject": "^4.0.0",
+    "gulp-istanbul": "^1.1.1",
+    "gulp-istanbul-enforcer": "^1.0.3",
+    "gulp-load-plugins": "^1.0.0-rc.1",
+    "gulp-mocha": "^2.1.3",
+    "gulp-node-inspector": "^0.1.0",
+    "gulp-plumber": "^1.0.1",
+    "gulp-protractor": "^3.0.0",
+    "gulp-rev": "^7.0.0",
+    "gulp-rev-replace": "^0.4.2",
+    "gulp-sort": "^2.0.0",
+    "gulp-sourcemaps": "^1.5.2",
+    "gulp-stylint": "^3.0.0",
+    "gulp-util": "^3.0.5",
+    "gulp-watch": "^4.3.5",
+    "html-webpack-harddisk-plugin": "~0.0.2",
+    "html-webpack-plugin": "^2.16.0",
+    "imports-loader": "^0.6.5",
+    "isparta": "^4.0.0",
+    "isparta-instrumenter-loader": "^1.0.0",
+    "isparta-loader": "^2.0.0",
+    "istanbul": "1.1.0-alpha.1",
+    "istanbul-instrumenter-loader": "^0.2.0",
     "jasmine-core": "^2.3.4",
-    "karma-jasmine": "~0.3.0",
     "jasmine-spec-reporter": "^2.4.0",
-    "karma-ng-scenario": "~0.1.0",
-    "karma-firefox-launcher": "~0.1.6",
-    "karma-script-launcher": "~0.1.0",
-    "karma-chrome-launcher": "~0.2.0",
-    "requirejs": "~2.1.11",
-    "karma-requirejs": "~0.2.2",
-    "karma-jade-preprocessor": "0.0.11",
-    "phantomjs": "^1.9.18",
-    "karma-phantomjs-launcher": "~0.2.0",
+    "jpegtran-bin": "^3.1.0",
     "karma": "~0.13.3",
-    "karma-ng-html2js-preprocessor": "~0.2.0",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-coverage": "^1.0.0",
+    "karma-firefox-launcher": "^1.0.0",
+    "karma-jasmine": "^1.0.2",
+    "karma-phantomjs-launcher": "~1.0.0",
+    "karma-script-launcher": "^1.0.0",
+    "karma-sourcemap-loader": "~0.3.7",
     "karma-spec-reporter": "~0.0.20",
+    "karma-webpack": "^1.7.0",
+    "lazypipe": "^1.0.1",
+    "mocha": "^3.0.2",
+    "ng-annotate-loader": "~0.1.0",
+    "node-sass": "^3.8.0",
+    "nodemon": "^1.3.7",
+    "null-loader": "^0.1.1",
+    "open": "~0.0.4",
+    "optipng-bin": "^3.1.2",
+    "phantomjs-prebuilt": "^2.1.4",
+    "postcss-loader": "^0.11.1",
     "proxyquire": "^1.0.1",
-    "supertest": "^1.1.0"
+    "pug-html-loader": "^1.0.8",
+    "raw-loader": "^0.5.1",
+    "run-sequence": "^1.1.0",
+    "sass-loader": "^4.0.1",
+    "sinon": "^1.16.1",
+    "sinon-chai": "^2.8.0",
+    "strip-ansi": "^3.0.1",
+    "style-loader": "^0.13.0",
+    "supertest": "^1.1.0",
+    "through2": "^2.0.1",
+    "webpack": "^1.12.14",
+    "webpack-dev-middleware": "^1.5.1",
+    "webpack-stream": "^3.2.0"
   },
   "engines": {
-    "node": "^4.2.3",
-    "npm": "^2.14.7"
+    "node": "^6.2.2",
+    "npm": "^3.9.5"
   },
   "scripts": {
-    "start": "node server",
-    "test": "grunt test",
-    "update-webdriver": "node node_modules/grunt-protractor-runner/node_modules/protractor/bin/webdriver-manager update"
+    "test": "gulp test",
+    "flow": "flow",
+    "update-webdriver": "node node_modules/protractor/bin/webdriver-manager update",
+    "start": "node server"
   },
   "private": true
 }
