@@ -1,14 +1,15 @@
 'use strict';
+/*eslint no-process-env:0*/
 
-var path = require('path');
-var _ = require('lodash');
+import path from 'path';
+import _ from 'lodash';
 
-function requiredProcessEnv(name) {
-  if (!process.env[name]) {
+/*function requiredProcessEnv(name) {
+  if(!process.env[name]) {
     throw new Error('You must set the ' + name + ' environment variable');
   }
   return process.env[name];
-}
+}*/
 
 // All configurations will extend these options
 // ============================================
@@ -16,7 +17,10 @@ var all = {
   env: process.env.NODE_ENV,
 
   // Root path of server
-  root: path.normalize(__dirname + '/../../..'),
+  root: path.normalize(`${__dirname}/../../..`),
+
+  // Browser-sync port
+  browserSyncPort: process.env.BROWSER_SYNC_PORT || 3000,
 
   // Server port
   port: process.env.PORT || 9000,
@@ -47,4 +51,4 @@ var all = {
 module.exports = _.merge(
   all,
   require('./shared'),
-  require('./' + process.env.NODE_ENV + '.js') || {});
+  require(`./${process.env.NODE_ENV}.js`) || {});
