@@ -7,8 +7,7 @@
 import errors from './components/errors';
 import path from 'path';
 
-module.exports = function(app) {
-
+export default function(app) {
   // Insert routes below
   app.use('/api/things', require('./api/thing'));
   // All undefined asset or api routes should return a 404
@@ -17,7 +16,7 @@ module.exports = function(app) {
 
   // All other routes should redirect to the index.html
   app.route('/*')
-    .get(function(req, res) {
-      res.sendFile(path.resolve(app.get('appPath') + '/index.html'));
+    .get((req, res) => {
+      res.sendFile(path.resolve(`${app.get('appPath')}/index.html`));
     });
-};
+}
