@@ -12,7 +12,7 @@ describe('Main View', function() {
 
   it('should include jumbotron with correct data', function() {
     expect(page.h1El.getText()).toBe('\'Allo, \'Allo!');
-    expect(page.imgEl.getAttribute('src')).toMatch(/yeoman.png$/);
+    expect(page.imgEl.getAttribute('src')).toMatch(/yeoman(\.[a-zA-Z0-9]*)?\.png$/);
     expect(page.imgEl.getAttribute('alt')).toBe('I\'m Yeoman');
   });
 });
