@@ -1,5 +1,8 @@
 // Karma configuration
-// http://karma-runner.github.io/0.10/config/configuration-file.html
+// http://karma-runner.github.io/0.13/config/configuration-file.html
+/*eslint-env node*/
+
+import makeWebpackConfig from './webpack.make';
 
 module.exports = function(config) {
   config.set({
@@ -10,54 +13,50 @@ module.exports = function(config) {
     frameworks: ['jasmine'],
 
     // list of files / patterns to load in the browser
-    files: [
-      // bower:js
-      'client/bower_components/jquery/dist/jquery.js',
-      'client/bower_components/angular/angular.js',
-      'client/bower_components/bootstrap/dist/js/bootstrap.js',
-      'client/bower_components/angular-resource/angular-resource.js',
-      'client/bower_components/angular-cookies/angular-cookies.js',
-      'client/bower_components/angular-sanitize/angular-sanitize.js',
-      'client/bower_components/angular-route/angular-route.js',
-      'client/bower_components/lodash/dist/lodash.compat.js',
-      'client/bower_components/ng-dialog/js/ngDialog.js',
-      'client/bower_components/angular-mocks/angular-mocks.js',
-      // endbower
-      'client/app/app.js',
-      'client/{app,components}/**/*.module.js',
-      'client/{app,components}/**/*.js',
-      'client/{app,components}/**/*.html'
-    ],
+    files: ['spec.js'],
 
     preprocessors: {
-      '**/*.html': 'ng-html2js',
-      'client/{app,components}/**/*.js': 'babel'
+      'spec.js': ['webpack']
     },
 
-    ngHtml2JsPreprocessor: {
-      stripPrefix: 'client/'
+    webpack: makeWebpackConfig({ TEST: true }),
+
+    webpackMiddleware: {
+      // webpack-dev-middleware configuration
+      // i. e.
+      noInfo: true
     },
 
-    babelPreprocessor: {
-      options: {
-        sourceMap: 'inline',
-        optional: [
-          'es7.classProperties'
-        ]
-      },
-      filename: function (file) {
-        return file.originalPath.replace(/\.js$/, '.es5.js');
-      },
-      sourceFileName: function (file) {
-        return file.originalPath;
-      }
+    coverageReporter: {
+      reporters: [{
+        type: 'html', //produces a html document after code is run
+        subdir: 'client'
+      }, {
+        type: 'json',
+        subdir: '.',
+        file: 'client-coverage.json'
+      }],
+      dir: 'coverage/' //path to created html doc
     },
+
+    plugins: [
+      require('karma-chrome-launcher'),
+      require('karma-coverage'),
+      require('karma-firefox-launcher'),
+
+      require('karma-jasmine'),
+      require('karma-spec-reporter'),
+      require('karma-phantomjs-launcher'),
+      require('karma-script-launcher'),
+      require('karma-webpack'),
+      require('karma-sourcemap-loader')
+    ],
 
     // list of files / patterns to exclude
     exclude: [],
 
     // web server port
-    port: 8080,
+    port: 9000,
 
     // level of logging
     // possible values: LOG_DISABLE || LOG_ERROR || LOG_WARN || LOG_INFO || LOG_DEBUG
@@ -70,7 +69,7 @@ module.exports = function(config) {
     // - junit
     // - growl
     // - coverage
-    reporters: ['spec'],
+    reporters: ['spec', 'coverage'],
 
     // enable / disable watching file and executing tests whenever any file changes
     autoWatch: false,
