@@ -59,16 +59,12 @@ var config = {
   },
 
   onPrepare: function() {
-    require('babel-core/register');
+    require('babel-register');
     var SpecReporter = require('jasmine-spec-reporter');
     // add jasmine spec reporter
     jasmine.getEnv().addReporter(new SpecReporter({displayStacktrace: true}));
 
     var serverConfig = config.params.serverConfig;
-
-    // Setup mongo for tests
-    var mongoose = require('mongoose');
-    mongoose.connect(serverConfig.mongo.uri, serverConfig.mongo.options); // Connect to database
   }
 };
 
