@@ -4,31 +4,19 @@
 
 'use strict';
 
-import newrelic from 'newrelic';
 import express from 'express';
-// import mongoose from 'mongoose';
 import config from './config/environment';
 import http from 'http';
-
-// Connect to MongoDB
-// mongoose.connect(config.mongo.uri, config.mongo.options);
-// mongoose.connection.on('error', function(err) {
-//   console.error('MongoDB connection error: ' + err);
-//   process.exit(-1);
-// });
-
-// Populate databases with sample data
-if (config.seedDB) { require('./config/seed'); }
 
 // Setup server
 var app = express();
 var server = http.createServer(app);
-require('./config/express')(app);
-require('./routes')(app);
+require('./config/express').default(app);
+require('./routes').default(app);
 
 // Start server
 function startServer() {
-  server.listen(config.port, config.ip, function() {
+  app.angularFullstack = server.listen(config.port, config.ip, function() {
     console.log('Express server listening on %d, in %s mode', config.port, app.get('env'));
   });
 }
