@@ -13,8 +13,8 @@ module.exports[404] = function pageNotFound(req, res) {
 
   res.status(result.status);
   res.render(viewFilePath, {}, function(err, html) {
-    if (err) {
-      return res.json(result, result.status);
+    if(err) {
+      return res.status(result.status).json(result);
     }
 
     res.send(html);
