@@ -1,15 +1,11 @@
 'use strict';
+/*eslint no-process-env:0*/
 
 // Development specific configuration
 // ==================================
 module.exports = {
 
-  // MongoDB connection options
-  // mongo: {
-  //   uri: 'mongodb://localhost/itinerateconsulting-dev'
-  // },
-
   // Seed database on startup
-  // seedDB: true
+  seedDB: true
 
 };
