@@ -1,24 +1,16 @@
 'use strict';
+/*eslint no-process-env:0*/
 
 // Production specific configuration
 // =================================
 module.exports = {
   // Server IP
-  ip:     process.env.OPENSHIFT_NODEJS_IP ||
-          process.env.IP ||
-          undefined,
+  ip: process.env.OPENSHIFT_NODEJS_IP
+    || process.env.ip
+    || undefined,
 
   // Server port
-  port:   process.env.OPENSHIFT_NODEJS_PORT ||
-          process.env.PORT ||
-          8080,
-
-  // // MongoDB connection options
-  // mongo: {
-  //   uri:  process.env.MONGOLAB_URI ||
-  //         process.env.MONGOHQ_URL ||
-  //         process.env.OPENSHIFT_MONGODB_DB_URL +
-  //         process.env.OPENSHIFT_APP_NAME ||
-  //         'mongodb://localhost/itinerateconsulting'
-  // }
+  port: process.env.OPENSHIFT_NODEJS_PORT
+    || process.env.PORT
+    || 8080
 };
