@@ -3,9 +3,9 @@
 // Set default node environment to development
 var env = process.env.NODE_ENV = process.env.NODE_ENV || 'development';
 
-if (env === 'development' || env === 'test') {
+if(env === 'development' || env === 'test') {
   // Register the Babel require hook
-  require('babel-core/register');
+  require('babel-register');
 }
 
 // Export the application
